requirements.txt
opencv-python==4.4.0.42
Pillow==7.2.0
tensorflow-gpu==2.3.0
tf-slim==1.1.0